LPS-166968 Simplify, inline the method getDataLayoutJSONObject as it is only used by one place
