+++ modules/apps/data-engine/data-engine-taglib/src/main/java/com/liferay/data/engine/taglib/internal/servlet/taglib/util/DataLayoutTaglibUtil.java
@@ -14,82 +14,37 @@
 
 package com.liferay.data.engine.taglib.internal.servlet.taglib.util;
 
-import com.liferay.data.engine.field.type.util.LocalizedValueUtil;
 import com.liferay.data.engine.rest.dto.v2_0.DataDefinition;
 import com.liferay.data.engine.rest.dto.v2_0.DataLayout;
-import com.liferay.data.engine.rest.dto.v2_0.DataRule;
-import com.liferay.data.engine.rest.dto.v2_0.util.DataDefinitionDDMFormUtil;
 import com.liferay.data.engine.rest.resource.v2_0.DataDefinitionResource;
 import com.liferay.data.engine.rest.resource.v2_0.DataLayoutResource;
-import com.liferay.data.engine.taglib.servlet.taglib.definition.DataLayoutBuilderDefinition;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldType;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeServicesRegistry;
-import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderingContext;
-import com.liferay.dynamic.data.mapping.form.renderer.DDMFormTemplateContextFactory;
-import com.liferay.dynamic.data.mapping.io.DDMFormLayoutDeserializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormLayoutDeserializerDeserializeRequest;
-import com.liferay.dynamic.data.mapping.io.DDMFormLayoutDeserializerDeserializeResponse;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormField;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidation;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidationExpression;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayoutColumn;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayoutPage;
-import com.liferay.dynamic.data.mapping.model.DDMFormLayoutRow;
-import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.DDMStructureLayout;
-import com.liferay.dynamic.data.mapping.model.LocalizedValue;
-import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
-import com.liferay.dynamic.data.mapping.model.Value;
-import com.liferay.dynamic.data.mapping.service.DDMStructureLayoutLocalService;
-import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
-import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
-import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
-import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
-import com.liferay.dynamic.data.mapping.util.DDMFormLayoutFactory;
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.string.StringPool;
+import com.liferay.osgi.util.service.Snapshot;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.LocaleThreadLocal;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceCardinality;
-import org.osgi.service.component.annotations.ReferencePolicy;
-import org.osgi.service.component.annotations.ReferencePolicyOption;
 
 /**
  * @author Gabriel Albuquerque
@@ -114,14 +69,24 @@ public class DataLayoutTaglibUtil {
 			dataDefinitionId, httpServletRequest);
 	}
 
-	public static JSONObject getDataLayoutJSONObject(
-		Set<Locale> availableLocales, String contentType, Long dataDefinitionId,
-		Long dataLayoutId, HttpServletRequest httpServletRequest,
-		HttpServletResponse httpServletResponse) {
+	public static DataLayout getDataLayout(
+			Long dataLayoutId, HttpServletRequest httpServletRequest)
+		throws Exception {
 
-		return _dataLayoutTaglibUtil._getDataLayoutJSONObject(
-			availableLocales, contentType, dataDefinitionId, dataLayoutId,
-			httpServletRequest, httpServletResponse);
+		DataLayoutResource.Factory dataLayoutResourceFactory =
+			_dataLayoutResourceFactorySnapshot.get();
+
+		DataLayoutResource.Builder dataLayoutResourceBuilder =
+			dataLayoutResourceFactory.create();
+
+		DataLayoutResource dataLayoutResource =
+			dataLayoutResourceBuilder.httpServletRequest(
+				httpServletRequest
+			).user(
+				PortalUtil.getUser(httpServletRequest)
+			).build();
+
+		return dataLayoutResource.getDataLayout(dataLayoutId);
 	}
 
 	public static Long getDefaultDataLayoutId(
@@ -146,43 +111,9 @@ public class DataLayoutTaglibUtil {
 		_dataLayoutTaglibUtil = this;
 	}
 
-	@Reference(
-		cardinality = ReferenceCardinality.MULTIPLE,
-		policy = ReferencePolicy.DYNAMIC,
-		policyOption = ReferencePolicyOption.GREEDY
-	)
-	protected void addDataLayoutBuilderDefinition(
-		DataLayoutBuilderDefinition dataLayoutBuilderDefinition,
-		Map<String, Object> properties) {
-
-		String contentType = GetterUtil.getString(
-			properties.get("content.type"));
-
-		if (Validator.isNull(contentType)) {
-			return;
-		}
-
-		_dataLayoutBuilderDefinitions.put(
-			contentType, dataLayoutBuilderDefinition);
-	}
-
 	@Deactivate
 	protected void deactivate() {
 		_dataLayoutTaglibUtil = null;
-	}
-
-	protected void removeDataLayoutBuilderDefinition(
-		DataLayoutBuilderDefinition dataLayoutBuilderDefinition,
-		Map<String, Object> properties) {
-
-		String contentType = GetterUtil.getString(
-			properties.get("content.type"));
-
-		if (Validator.isNull(contentType)) {
-			return;
-		}
-
-		_dataLayoutBuilderDefinitions.remove(contentType);
 	}
 
 	private Set<Locale> _getAvailableLocales(
@@ -205,7 +136,7 @@ public class DataLayoutTaglibUtil {
 					dataDefinitionId, httpServletRequest);
 			}
 			else {
-				DataLayout dataLayout = _getDataLayout(
+				DataLayout dataLayout = getDataLayout(
 					dataLayoutId, httpServletRequest);
 
 				dataDefinition = _getDataDefinition(
@@ -242,83 +173,6 @@ public class DataLayoutTaglibUtil {
 			).build();
 
 		return dataDefinitionResource.getDataDefinition(dataDefinitionId);
-	}
-
-	private DataLayout _getDataLayout(
-			Long dataLayoutId, HttpServletRequest httpServletRequest)
-		throws Exception {
-
-		DataLayoutResource.Builder dataLayoutResourceBuilder =
-			_dataLayoutResourceFactory.create();
-
-		DataLayoutResource dataLayoutResource =
-			dataLayoutResourceBuilder.httpServletRequest(
-				httpServletRequest
-			).user(
-				_portal.getUser(httpServletRequest)
-			).build();
-
-		return dataLayoutResource.getDataLayout(dataLayoutId);
-	}
-
-	private DataLayoutBuilderDefinition _getDataLayoutBuilderDefinition(
-		String contentType) {
-
-		DataLayoutBuilderDefinition dataLayoutBuilderDefinition =
-			_dataLayoutBuilderDefinitions.get(contentType);
-
-		if (dataLayoutBuilderDefinition == null) {
-			return _dataLayoutBuilderDefinitions.get("default");
-		}
-
-		return dataLayoutBuilderDefinition;
-	}
-
-	private JSONObject _getDataLayoutJSONObject(
-		Set<Locale> availableLocales, String contentType, Long dataDefinitionId,
-		Long dataLayoutId, HttpServletRequest httpServletRequest,
-		HttpServletResponse httpServletResponse) {
-
-		try {
-			String dataLayoutString = ParamUtil.getString(
-				httpServletRequest, "dataLayout");
-
-			if (Validator.isNotNull(dataLayoutString)) {
-				DataLayoutDDMFormAdapter dataLayoutDDMFormAdapter =
-					new DataLayoutDDMFormAdapter(
-						availableLocales, contentType,
-						DataLayout.toDTO(dataLayoutString), httpServletRequest,
-						httpServletResponse);
-
-				return dataLayoutDDMFormAdapter.toJSONObject();
-			}
-
-			DataLayout dataLayout = null;
-
-			if (Validator.isNotNull(dataLayoutId)) {
-				dataLayout = _getDataLayout(dataLayoutId, httpServletRequest);
-			}
-			else {
-				DataDefinition dataDefinition = _getDataDefinition(
-					dataDefinitionId, httpServletRequest);
-
-				dataLayout = dataDefinition.getDefaultDataLayout();
-			}
-
-			DataLayoutDDMFormAdapter dataLayoutDDMFormAdapter =
-				new DataLayoutDDMFormAdapter(
-					availableLocales, contentType, dataLayout,
-					httpServletRequest, httpServletResponse);
-
-			return dataLayoutDDMFormAdapter.toJSONObject();
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-
-			return _jsonFactory.createJSONObject();
-		}
 	}
 
 	private Long _getDefaultDataLayoutId(
@@ -425,492 +279,18 @@ public class DataLayoutTaglibUtil {
 	private static final Log _log = LogFactoryUtil.getLog(
 		DataLayoutTaglibUtil.class);
 
-	private static final Map<String, DataLayoutBuilderDefinition>
-		_dataLayoutBuilderDefinitions = new ConcurrentHashMap<>();
+	private static final Snapshot<DataLayoutResource.Factory>
+		_dataLayoutResourceFactorySnapshot = new Snapshot<>(
+			DataLayoutTaglibUtil.class, DataLayoutResource.Factory.class);
 	private static DataLayoutTaglibUtil _dataLayoutTaglibUtil;
 
 	@Reference
 	private DataDefinitionResource.Factory _dataDefinitionResourceFactory;
 
 	@Reference
-	private DataLayoutResource.Factory _dataLayoutResourceFactory;
-
-	@Reference
-	private DDMFormFieldTypeServicesRegistry _ddmFormFieldTypeServicesRegistry;
-
-	@Reference
-	private DDMFormTemplateContextFactory _ddmFormTemplateContextFactory;
-
-	@Reference
-	private DDMStructureLayoutLocalService _ddmStructureLayoutLocalService;
-
-	@Reference
-	private DDMStructureLocalService _ddmStructureLocalService;
-
-	@Reference(target = "(ddm.form.layout.deserializer.type=json)")
-	private DDMFormLayoutDeserializer _jsonDDMFormLayoutDeserializer;
-
-	@Reference
 	private JSONFactory _jsonFactory;
 
 	@Reference
-	private Language _language;
-
-	@Reference
 	private Portal _portal;
-
-	private class DataLayoutDDMFormAdapter {
-
-		public DataLayoutDDMFormAdapter(
-			Set<Locale> availableLocales, String contentType,
-			DataLayout dataLayout, HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse) {
-
-			_availableLocales = availableLocales;
-			_contentType = contentType;
-			_dataLayout = dataLayout;
-			_httpServletRequest = httpServletRequest;
-			_httpServletResponse = httpServletResponse;
-		}
-
-		public JSONObject toJSONObject() throws Exception {
-			DDMForm ddmForm = null;
-
-			if (_dataLayout.getId() == null) {
-				DataDefinition dataDefinition = DataDefinition.toDTO(
-					_httpServletRequest.getParameter("dataDefinition"));
-
-				ddmForm = DataDefinitionDDMFormUtil.toDDMForm(
-					dataDefinition, _ddmFormFieldTypeServicesRegistry);
-			}
-			else {
-				ddmForm = _getDDMForm();
-			}
-
-			Locale defaultLocale = ddmForm.getDefaultLocale();
-
-			Map<String, Object> ddmFormTemplateContext =
-				_ddmFormTemplateContextFactory.create(
-					ddmForm, _getDDMFormLayout(),
-					new DDMFormRenderingContext() {
-						{
-							setHttpServletRequest(_httpServletRequest);
-							setHttpServletResponse(_httpServletResponse);
-							setLocale(defaultLocale);
-							setPortletNamespace(StringPool.BLANK);
-						}
-					});
-
-			_populateDDMFormFieldSettingsContext(
-				ddmForm.getDDMFormFieldsMap(true), ddmFormTemplateContext,
-				defaultLocale);
-
-			ddmFormTemplateContext.put("rules", _getDataRulesJSONArray());
-
-			return _jsonFactory.createJSONObject(
-				_jsonFactory.looseSerializeDeep(ddmFormTemplateContext));
-		}
-
-		private Map<String, Object> _createDDMFormFieldSettingContext(
-				DDMFormField ddmFormField, Locale defaultLocale)
-			throws Exception {
-
-			DDMFormFieldType ddmFormFieldType =
-				_ddmFormFieldTypeServicesRegistry.getDDMFormFieldType(
-					ddmFormField.getType());
-
-			DDMForm ddmForm = DDMFormFactory.create(
-				ddmFormFieldType.getDDMFormFieldTypeSettings());
-
-			DDMFormLayout ddmFormLayout = DDMFormLayoutFactory.create(
-				ddmFormFieldType.getDDMFormFieldTypeSettings());
-
-			_removeDisabledProperties(ddmForm, ddmFormLayout);
-
-			return _ddmFormTemplateContextFactory.create(
-				ddmForm, ddmFormLayout,
-				new DDMFormRenderingContext() {
-					{
-						setContainerId("settings");
-						setDDMFormValues(
-							_createDDMFormFieldSettingContextDDMFormValues(
-								ddmFormField, ddmForm));
-						setHttpServletRequest(_httpServletRequest);
-						setHttpServletResponse(_httpServletResponse);
-						setLocale(defaultLocale);
-						setPortletNamespace(StringPool.BLANK);
-					}
-				});
-		}
-
-		private DDMFormValues _createDDMFormFieldSettingContextDDMFormValues(
-				DDMFormField ddmFormField,
-				DDMForm ddmFormFieldTypeSettingsDDMForm)
-			throws Exception {
-
-			DDMFormValues ddmFormValues = new DDMFormValues(
-				ddmFormFieldTypeSettingsDDMForm);
-
-			DDMForm ddmForm = ddmFormField.getDDMForm();
-			Map<String, Object> ddmFormFieldProperties =
-				ddmFormField.getProperties();
-
-			for (DDMFormField ddmFormFieldTypeSetting :
-					ddmFormFieldTypeSettingsDDMForm.getDDMFormFields()) {
-
-				ddmFormValues.addDDMFormFieldValue(
-					new DDMFormFieldValue() {
-						{
-							setName(ddmFormFieldTypeSetting.getName());
-							setValue(
-								_createDDMFormFieldValue(
-									ddmForm.getAvailableLocales(),
-									ddmFormFieldTypeSetting,
-									ddmFormFieldProperties.get(
-										ddmFormFieldTypeSetting.getName())));
-						}
-					});
-			}
-
-			return ddmFormValues;
-		}
-
-		private Value _createDDMFormFieldValue(
-			DDMFormFieldValidation ddmFormFieldValidation) {
-
-			if (ddmFormFieldValidation == null) {
-				return new UnlocalizedValue(StringPool.BLANK);
-			}
-
-			DDMFormFieldValidationExpression ddmFormFieldValidationExpression =
-				ddmFormFieldValidation.getDDMFormFieldValidationExpression();
-
-			return new UnlocalizedValue(
-				JSONUtil.put(
-					"errorMessage",
-					LocalizedValueUtil.toJSONObject(
-						LocalizedValueUtil.toLocalizedValuesMap(
-							ddmFormFieldValidation.
-								getErrorMessageLocalizedValue()))
-				).put(
-					"expression",
-					JSONUtil.put(
-						"name", ddmFormFieldValidationExpression.getName()
-					).put(
-						"value", ddmFormFieldValidationExpression.getValue()
-					)
-				).put(
-					"parameter",
-					LocalizedValueUtil.toJSONObject(
-						LocalizedValueUtil.toLocalizedValuesMap(
-							ddmFormFieldValidation.
-								getParameterLocalizedValue()))
-				).toString());
-		}
-
-		private Value _createDDMFormFieldValue(
-				Set<Locale> availableLocales,
-				DDMFormField ddmFormFieldTypeSetting, Object propertyValue)
-			throws Exception {
-
-			if (ddmFormFieldTypeSetting.isLocalizable()) {
-				return (LocalizedValue)propertyValue;
-			}
-
-			if (Objects.equals(
-					ddmFormFieldTypeSetting.getDataType(), "ddm-options")) {
-
-				if (propertyValue == null) {
-					propertyValue = new DDMFormFieldOptions();
-				}
-
-				return _createDDMFormFieldValue(
-					availableLocales, (DDMFormFieldOptions)propertyValue);
-			}
-
-			if (Objects.equals(
-					ddmFormFieldTypeSetting.getName(), "requiredDescription") &&
-				(propertyValue == null)) {
-
-				return new UnlocalizedValue(Boolean.TRUE.toString());
-			}
-
-			if (Objects.equals(
-					ddmFormFieldTypeSetting.getType(), "validation")) {
-
-				return _createDDMFormFieldValue(
-					(DDMFormFieldValidation)propertyValue);
-			}
-
-			return new UnlocalizedValue(String.valueOf(propertyValue));
-		}
-
-		private Value _createDDMFormFieldValue(
-				Set<Locale> availableLocales,
-				DDMFormFieldOptions ddmFormFieldOptions)
-			throws Exception {
-
-			JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-			for (Locale availableLocale : availableLocales) {
-				jsonObject.put(
-					LocaleUtil.toLanguageId(availableLocale),
-					JSONUtil.toJSONArray(
-						ddmFormFieldOptions.getOptionsValues(),
-						optionValue -> {
-							LocalizedValue localizedValue =
-								ddmFormFieldOptions.getOptionLabels(
-									optionValue);
-
-							return JSONUtil.put(
-								"label",
-								localizedValue.getString(availableLocale)
-							).put(
-								"reference",
-								ddmFormFieldOptions.getOptionReference(
-									optionValue)
-							).put(
-								"value", optionValue
-							);
-						}));
-			}
-
-			return new UnlocalizedValue(jsonObject.toString());
-		}
-
-		private DDMFormLayout _deserializeDDMFormLayout(String content) {
-			DDMFormLayoutDeserializerDeserializeRequest.Builder builder =
-				DDMFormLayoutDeserializerDeserializeRequest.Builder.newBuilder(
-					content);
-
-			DDMFormLayoutDeserializerDeserializeResponse
-				ddmFormLayoutDeserializerDeserializeResponse =
-					_jsonDDMFormLayoutDeserializer.deserialize(builder.build());
-
-			return ddmFormLayoutDeserializerDeserializeResponse.
-				getDDMFormLayout();
-		}
-
-		private JSONArray _getDataRulesJSONArray() {
-			JSONArray dataRulesJSONArray = _jsonFactory.createJSONArray();
-
-			for (DataRule dataRule : _dataLayout.getDataRules()) {
-				JSONObject dataRuleJSONObject = _jsonFactory.createJSONObject();
-
-				JSONArray jsonArray = _jsonFactory.createJSONArray();
-
-				for (Map<String, Object> action : dataRule.getActions()) {
-					JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-					action.forEach(jsonObject::put);
-
-					jsonArray.put(jsonObject);
-				}
-
-				dataRuleJSONObject.put("actions", jsonArray);
-
-				jsonArray = _jsonFactory.createJSONArray();
-
-				for (Map<String, Object> condition : dataRule.getConditions()) {
-					JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-					condition.forEach(jsonObject::put);
-
-					jsonArray.put(jsonObject);
-				}
-
-				dataRuleJSONObject.put(
-					"conditions", jsonArray
-				).put(
-					"logicalOperator", dataRule.getLogicalOperator()
-				).put(
-					"name", LocalizedValueUtil.toJSONObject(dataRule.getName())
-				);
-
-				dataRulesJSONArray.put(dataRuleJSONObject);
-			}
-
-			return dataRulesJSONArray;
-		}
-
-		private DDMForm _getDDMForm() throws Exception {
-			DDMStructure ddmStructure = _ddmStructureLocalService.getStructure(
-				_dataLayout.getDataDefinitionId());
-
-			String dataDefinitionJSON = ddmStructure.getDefinition();
-
-			JSONObject jsonObject = _jsonFactory.createJSONObject(
-				StringUtil.replace(
-					dataDefinitionJSON, "defaultValue", "predefinedValue"));
-
-			ddmStructure.setDefinition(
-				jsonObject.put(
-					"availableLanguageIds",
-					JSONUtil.toJSONArray(
-						_availableLocales,
-						availableLocale -> _language.getLanguageId(
-							availableLocale))
-				).put(
-					"defaultLanguageId", ddmStructure.getDefaultLanguageId()
-				).toString());
-
-			return ddmStructure.getDDMForm();
-		}
-
-		private DDMFormLayout _getDDMFormLayout() throws Exception {
-			String definition = null;
-
-			if (_dataLayout.getId() == null) {
-				definition = _dataLayout.toString();
-			}
-			else {
-				DDMStructureLayout ddmStructureLayout =
-					_ddmStructureLayoutLocalService.getStructureLayout(
-						_dataLayout.getId());
-
-				definition = ddmStructureLayout.getDefinition();
-			}
-
-			JSONObject jsonObject = _jsonFactory.createJSONObject(
-				StringUtil.replace(
-					definition,
-					new String[] {
-						"columnSize", "dataLayoutColumns", "dataLayoutPages",
-						"dataLayoutRows"
-					},
-					new String[] {"size", "columns", "pages", "rows"}));
-
-			return _deserializeDDMFormLayout(jsonObject.toString());
-		}
-
-		private List<Map<String, Object>> _getNestedFields(
-			Map<String, Object> field) {
-
-			List<Map<String, Object>> nestedFields = new ArrayList<>();
-
-			List<Map<String, Object>> fieldNestedFields =
-				(List<Map<String, Object>>)field.get("nestedFields");
-
-			if (fieldNestedFields == null) {
-				return nestedFields;
-			}
-
-			for (Map<String, Object> nestedField : fieldNestedFields) {
-				nestedFields.add(nestedField);
-
-				nestedFields.addAll(_getNestedFields(nestedField));
-			}
-
-			return nestedFields;
-		}
-
-		private boolean _isFieldSet(Map<String, Object> field) {
-			if (Objects.equals(field.get("type"), "fieldset")) {
-				return true;
-			}
-
-			return false;
-		}
-
-		private void _populateDDMFormFieldSettingsContext(
-				Map<String, DDMFormField> ddmFormFieldsMap,
-				Map<String, Object> ddmFormTemplateContext,
-				Locale defaultLocale)
-			throws Exception {
-
-			UnsafeConsumer<Map<String, Object>, Exception> unsafeConsumer =
-				field -> {
-					DDMFormField ddmFormField = ddmFormFieldsMap.get(
-						MapUtil.getString(field, "fieldName"));
-
-					if (_isFieldSet(field)) {
-						ddmFormField.setProperty("rows", field.get("rows"));
-					}
-
-					field.put(
-						"settingsContext",
-						_createDDMFormFieldSettingContext(
-							ddmFormField, defaultLocale));
-				};
-
-			List<Map<String, Object>> pages =
-				(List<Map<String, Object>>)ddmFormTemplateContext.get("pages");
-
-			for (Map<String, Object> page : pages) {
-				List<Map<String, Object>> rows =
-					(List<Map<String, Object>>)page.get("rows");
-
-				for (Map<String, Object> row : rows) {
-					List<Map<String, Object>> columns =
-						(List<Map<String, Object>>)row.get("columns");
-
-					for (Map<String, Object> column : columns) {
-						List<Map<String, Object>> fields =
-							(List<Map<String, Object>>)column.get("fields");
-
-						for (Map<String, Object> field : fields) {
-							unsafeConsumer.accept(field);
-
-							List<Map<String, Object>> nestedFields =
-								_getNestedFields(field);
-
-							for (Map<String, Object> nestedField :
-									nestedFields) {
-
-								unsafeConsumer.accept(nestedField);
-							}
-						}
-					}
-				}
-			}
-		}
-
-		private void _removeDisabledProperties(
-			DDMForm ddmForm, DDMFormLayout ddmFormLayout) {
-
-			DataLayoutBuilderDefinition dataLayoutBuilderDefinition =
-				_getDataLayoutBuilderDefinition(_contentType);
-
-			String[] disabledProperties =
-				dataLayoutBuilderDefinition.getDisabledProperties();
-
-			if (ArrayUtil.isEmpty(disabledProperties)) {
-				return;
-			}
-
-			for (String disabledProperty : disabledProperties) {
-				Map<String, DDMFormField> ddmFormFieldsMap =
-					ddmForm.getDDMFormFieldsMap(true);
-
-				List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
-
-				ddmFormFields.remove(ddmFormFieldsMap.get(disabledProperty));
-
-				for (DDMFormLayoutPage ddmFormLayoutPage :
-						ddmFormLayout.getDDMFormLayoutPages()) {
-
-					for (DDMFormLayoutRow ddmFormLayoutRow :
-							ddmFormLayoutPage.getDDMFormLayoutRows()) {
-
-						for (DDMFormLayoutColumn ddmFormLayoutColumn :
-								ddmFormLayoutRow.getDDMFormLayoutColumns()) {
-
-							List<String> ddmFormFieldNames =
-								ddmFormLayoutColumn.getDDMFormFieldNames();
-
-							ddmFormFieldNames.remove(disabledProperty);
-						}
-					}
-				}
-			}
-		}
-
-		private final Set<Locale> _availableLocales;
-		private final String _contentType;
-		private final DataLayout _dataLayout;
-		private final HttpServletRequest _httpServletRequest;
-		private final HttpServletResponse _httpServletResponse;
-
-	}
 
 }
